Refactor actual dump / backup commands to standalone helper methods
Rather than compiling the commands from variables inside the template,
move them to a standalone helper library of their own. This allows
them to be much more robustly specified, so that we can be clearer
about the logic behind each separate command rather than having to
parse the full template for each set of variables.